Проведены исправления ошибок

+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -3,15 +3,12 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import ru.skypro.homework.entity.Ad;
 import ru.skypro.homework.entity.User;
-import java.util.Optional;
-import java.io.IOException;
 
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     @Query("SELECT s from Ad s")
     List<User> getAll();
     User findUserByUsername(String username);
